Listener now uses scoped lifetime (#40)
:warning: Breaking Changes

NRuuviTagServiceCollectionExtensions now registers the IRuuviTagListener as a scoped service instead of transient. This is a breaking change but makes no practical difference in the CLI since a new scope is already created for the Spectre command app.

The registration of the HTTP client for the HTTP publisher has been moved to the NRuuviTagServiceCollectionExtensions.AddCoreRuuviTagServices method.