Listener refactor for data format 6 and options pattern
Refactors IRuuviTagListener and RuuviTagListener and implementations so that support for data format 6 can be enabled on systems that do not support extended advertisements.

IRuuviTagListener.ListenAsync no longer has an overload that allows a filter to be specified - all filtering is now handled within the listener implementation.

RuuviTagListener now uses an options object to define listener options, and exposes an IDeviceResolver to implementations to allow filtering of unknown devices at the listener level.