Add support for data format 6 and refactor listener and producer base classes (#43)
* Add support for data format 6

Data format 6 is an alternative to Extended v1 for receivers that do not support extended advertisements

* Add new IDeviceLookup service for resolving devices

* Rename IDeviceLookup to IDeviceResolver

* Listener refactor for data format 6 and options pattern

Refactors IRuuviTagListener and RuuviTagListener and implementations so that support for data format 6 can be enabled on systems that do not support extended advertisements.

IRuuviTagListener.ListenAsync no longer has an overload that allows a filter to be specified - all filtering is now handled within the listener implementation.

RuuviTagListener now uses an options object to define listener options, and exposes an IDeviceResolver to implementations to allow filtering of unknown devices at the listener level.

* Publishers no longer supply filter delegates

RuuviTagPublisher implementations no longer supply a MAC address filter delegate - filtering of samples is now primarily the responsibility of the underlying listener.

All publishers can (through their options) specify a PrepareForPublish delegate, which allows modification or filtering of samples on a case-by-case basis if required.

* Add listener factory to CLI project

IRuuviTagListenerFactory is a new service for creating IRuuviTagListener instances. This allows listener features (such as enabling data format 6) to configured via the command app instead of being defined up-front when the DI container is being created.

* Use listener factory and consolidate common settings

* Update unit tests

* Update sample app

* Fix incorrect service registration

* Fix incorrect data format 6 switch

* Base listener now handles parsing logic

Implementations just need to notify the base listener about new advertisements and the base takes care of filtering and parsing.

* Update main README